STOPSHIP: vdso: disable switches for experiments
Permit vdso to be enabled or disabled at will to manage performance
experiments on the dogfood population.  Parameters are accessible
from user space at /sys/module/vdso/parameters/enable_{32|64}:

    enable_64=0 -> 64 bit vdso disabled
    enable_32=0 -> 32 bit vdso disabled

Overhead appears to be ~2ns to perform the checking on every call.
This is observed to give us a ~.4% screen on power increase

Change-Id: I2f0d9ae8ef8942ac43df502949453de40f4199f1
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Signed-off-by: Miguel de Dios <migueldedios@google.com>
Test: manual, bionic-benchmarks --bionic_xml=vdso.xml to confirm.
Bug: 70518189
Bug: 119630636
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: zclkkk <2484392714@qq.com>